Extend the table to include median
The median in addition to mean and 0.01th percentile as the lowest end
of IOPS/bw in addition to 99.99th are now included.

The median and the average/mean are usually close enough, but when it
comes to peaky SSDs (fast cache in front of slow backend SSD cells),
knowing the median (50th percentile) is necessary.